requirements.txt
requests
python-dotenv
beautifulsoup4
lxml
readability-lxml
lxml-html-clean